requirements.txt
Flask
opencv-python
pytesseract
Pillow
ebaysdk
numpy
python-dotenv
easyocr
